Improve UX of SublimeKSP menu actions (#129)
Instead of listing all the menu entries in a flat list, they are now tucked in a submenu at the end of Tools menu of SublimeKSP. This is also consistent with how some other Sublime Text plugins work. Also renamed the captions so that they don't unnecessarily repeat "KSP:" all the time.