Don't expand empty strings from an inline string array declaration (#184)
* Don't expand empty strings from an inline string array declaration

Up until now, SublimeKSP just expanded any and all strings from an inline string array declaration, which could result in a lot of unnecessary empty string assignments with larger (especially multidimensional) string array inline declarations.

Be smarter about it by doing a string value check, if the placeholder points to `""` then just don't append that string to `newLines`, which cleans up the compiled code.

* Make sure we're not redoing the placeholder check on already resolved placeholders

When importing files, inline initialized string array placeholderes are already resolved. Initial commit of this PR would end up with a traceback and not finish compilation as a consequence.

This commit fixes that.

* Make the placeholder check a regular expression

* Actually it was the indentation that was wrong, fix it

* Change how strVal is recreated from the placeholder

It is possible that sometimes a string define is mixed up with an inline string declaration, and this case was not handled properly before. Example:

`define TEST := "foo"
declare !str[] := ("banana", TEST & "bar")`

Now it works.