Improved attempted recompilation during existing compilation handling
If user attamptes to recompile the same script as the last one while we have a running compilation,
it is assumed there's been a quick update, so we will abort current compilation and start a new one.

If user attempts to recompile a different script from the last one,
show a message to wait until the current compilation is done.

Also move the "Compilation aborted!" message to the relevant method in ksp_compiler.py.